project_assets/tests/embeddings/requirements.txt
pandas
transformers
sentence-transformers
open-clip-torch
pillow
faiss-gpu
numpy==1.26.4